fix(review): keep recovery helpers within package guard
